Additional UI and other fixes (#24)
* Prompt cleanup, and other cleanup

* UI updates, moved a bunch of stuff to shared streamlit py file.
