+++ pages/1_General_AI.py
@@ -14,7 +14,6 @@ from src.ai.request_router import RequestRouter
 from src.configuration.assistant_configuration import AssistantConfigurationLoader
 
 
-
 from src.db.models.interactions import Interactions
 from src.db.models.documents import Documents
 from src.db.models.users import Users
@@ -23,53 +22,24 @@ from src.db.models.domain.document_model import DocumentModel
 
 from src.documents.document_loader import load_and_split_documents
 
-#from src.runners.ui.streamlit_agent_callback import StreamlitAgentCallbackHandler
+# from src.runners.ui.streamlit_agent_callback import StreamlitAgentCallbackHandler
 
 from src.ai.llm_helper import get_prompt
 from src.utilities.hash_utilities import calculate_sha256
 from src.ai.callbacks.streamlit_callbacks import StreamingOnlyCallbackHandler
 
+import src.ui.streamlit_shared as ui_shared
+
+
 class GeneralUI:
+    def __init__(self):
+        self.user_email = None
+
     def get_configuration_path(self):
         return os.environ.get(
             "ASSISTANT_CONFIG_PATH",
             "configurations/console_configs/console_ai.json",
         )
-
-    def get_available_collections(self, interaction_id) -> dict[str, int]:
-        documents_helper = Documents()
-
-        collections = documents_helper.get_collections(interaction_id)
-
-        # Create a dictionary of collection id to collection summary
-        collections_dict = {
-            collection.collection_name: collection.id for collection in collections
-        }
-
-        return collections_dict
-
-    def collection_id_from_option(self, option, interaction_id):
-        collections_dict = self.get_available_collections(interaction_id)
-
-        if option in collections_dict:
-            return collections_dict[option]
-        else:
-            return None
-
-    def create_collection(self, name):
-        selected_interaction_id = self.get_selected_interaction_id()
-
-        print(f"Creating collection {name} (interaction id: {selected_interaction_id})")
-
-        documents_helper = Documents()
-        collection = documents_helper.create_collection(
-            name,
-            selected_interaction_id,
-        )
-
-        print(f"Created collection {collection.collection_name}")
-
-        return collection.id
 
     def create_collections_container(self, main_window_container):
         css_style = """{
@@ -84,7 +54,7 @@ class GeneralUI:
     overflow-x: hidden;   /* Hides horizontal scrollbar */
 }"""
 
-        selected_interaction_id = self.get_selected_interaction_id()
+        selected_interaction_id = ui_shared.get_selected_interaction_id()
 
         with main_window_container:
             with stylable_container(key="collections_container", css_styles=css_style):
@@ -94,7 +64,7 @@ class GeneralUI:
                     col1, col2 = st.columns([0.80, 0.2])
                     col1.selectbox(
                         "Active document collection",
-                        self.get_available_collections(selected_interaction_id),
+                        ui_shared.get_available_collections(selected_interaction_id),
                         key="active_collection",
                         label_visibility="collapsed",
                     )
@@ -114,7 +84,7 @@ class GeneralUI:
                             st.session_state.get("new_collection_name")
                             and new_collection
                         ):
-                            self.create_collection(
+                            ui_shared.create_collection(
                                 st.session_state["new_collection_name"]
                             )
                             st.experimental_rerun()
@@ -122,7 +92,7 @@ class GeneralUI:
                     if "general_ai" in st.session_state:
                         option = st.session_state["active_collection"]
                         if option:
-                            collection_id = self.collection_id_from_option(
+                            collection_id = ui_shared.collection_id_from_option(
                                 option, selected_interaction_id
                             )
 
@@ -150,69 +120,23 @@ class GeneralUI:
                         else:
                             st.warning("No collection selected")
 
-    def get_interaction_pairs(self):
-        """Gets the interactions for the current user in 'UUID:STR' format"""
-        interactions_helper = Interactions()
-
-        interactions = None
-
-        user_id = self.set_user_id_from_email()
-
-        interactions = interactions_helper.get_interactions_by_user_id(user_id)
-
-        if not interactions:
-            return None
-
-        # Reverse the list so the most recent interactions are at the top
-        interactions.reverse()
-
-        interaction_pairs = [f"{i.id}:{i.interaction_summary}" for i in interactions]
-
-        print(f"get_interaction_pairs: interaction_pairs: {str(interaction_pairs)}")
-
-        return interaction_pairs
-
-    def load_interaction_selectbox(self):
-        """Loads the interaction selectbox"""
-
-        try:
-            st.sidebar.selectbox(
-                "Select Conversation",
-                self.get_interaction_pairs(),
-                key="interaction_summary_selectbox",
-                format_func=lambda x: x.split(":")[1],
-                on_change=self.load_ai,
-            )
-        except Exception as e:
-            print(f"Error loading interaction selectbox: {e}")
-
-    def get_selected_interaction_id(self):
-        """Gets the selected interaction id from the selectbox"""
-        selected_interaction_pair = st.session_state.get(
-            "interaction_summary_selectbox"
-        )
-
-        if not selected_interaction_pair:
-            return None
-
-        selected_interaction_id = selected_interaction_pair.split(":")[0]
-
-        print(
-            f"get_selected_interaction_id: selected_interaction_id: {selected_interaction_id}"
-        )
-
-        return selected_interaction_id
+            # st.write("")
+            # st.write("")
+            # st.write("")
+            # st.write("")
+            # st.write("")
+            # st.write("")
 
     def load_ai(self):
         """Loads the AI instance for the selected interaction id"""
-        selected_interaction_id = self.get_selected_interaction_id()
+        selected_interaction_id = ui_shared.get_selected_interaction_id()
 
         if "general_ai" not in st.session_state:
             # First time loading the page
             print("load_ai: ai not in session state")
             general_ai_instance = RequestRouter(
                 st.session_state["general_config"],
-                user_email,
+                self.user_email,
                 selected_interaction_id,
                 streaming=True,
             )
@@ -227,295 +151,11 @@ class GeneralUI:
             )
             general_ai_instance = RequestRouter(
                 st.session_state["general_config"],
-                user_email,
+                self.user_email,
                 selected_interaction_id,
                 streaming=True,
             )
             st.session_state["general_ai"] = general_ai_instance
-
-    def select_conversation(self):
-        with st.sidebar.container():
-            new_chat_button_clicked = st.sidebar.button(
-                "New Chat", key="new_chat_button"
-            )
-
-            if new_chat_button_clicked:
-                self.create_interaction("Empty Chat")
-                st.experimental_rerun()
-
-    def select_documents(self):
-        with st.sidebar.container():
-            st.toggle("Show LLM thoughts", key="show_llm_thoughts", value=True)
-
-            active_collection = st.session_state.get("active_collection")
-
-            uploaded_files = st.file_uploader(
-                "Choose your files",
-                accept_multiple_files=True,
-                disabled=(active_collection == None),
-                key="file_uploader",
-            )
-
-            status = st.status(f"File status", expanded=False, state="complete")
-
-            if uploaded_files and active_collection:
-                collection_id = None
-
-                if active_collection:
-                    collection_id = self.collection_id_from_option(
-                        active_collection,
-                        st.session_state["general_ai"].interaction_manager.interaction_id,
-                    )
-                    print(f"Active collection: {active_collection}")
-
-                with st.expander("File Ingestion Options", expanded=False):
-                    st.toggle(
-                        "Upload files into same dir (important for code)",
-                        key="use_same_upload_dir",
-                        value=True,
-                    )
-                    st.toggle(
-                        "Overwrite existing files",
-                        key="overwrite_existing_files",
-                        value=True,
-                    )
-                    st.toggle("Split documents", key="split_documents", value=True)
-                    st.text_input("Chunk size", key="file_chunk_size", value=500)
-                    st.text_input("Chunk overlap", key="file_chunk_overlap", value=50)
-
-                if (
-                    active_collection
-                    and st.button("Ingest files")
-                    and len(uploaded_files) > 0
-                ):
-                    self.ingest_files(
-                        uploaded_files,
-                        active_collection,
-                        collection_id,
-                        status,
-                        st.session_state["overwrite_existing_files"],
-                        st.session_state["split_documents"],
-                        int(st.session_state.get("file_chunk_size", 500)),
-                        int(st.session_state.get("file_chunk_overlap", 50)),
-                    )
-
-    def clear_to_create_file_and_docs(self, file_name, file_hash, collection_id, overwrite_existing_files) -> bool:
-        documents_helper = Documents()
-
-        existing_file = documents_helper.get_file_by_name(file_name, collection_id)        
-            
-        if existing_file is not None:
-            # File already exists
-
-            if overwrite_existing_files:
-                logging.info(
-                    f"Attempting to overwrite file: {file_name} in collection: {collection_id}"
-                )
-
-                # Did we find a file, and does it have a different hash?
-                if existing_file.file_hash != file_hash:
-                    # Delete all of the document chunks associated with this file
-                    documents_helper.delete_document_chunks_by_file_id(existing_file.id)
-
-                    # Delete the file itself
-                    documents_helper.delete_file(existing_file.id)
-                else:
-                    logging.info(
-                        f"Skipping overwrite- File: {file_name} in collection: {collection_id} already exists, and has the same hash"
-                    )
-                    st.info(f"Skipping overwrite- File: {file_name} in collection: {collection_id} already exists, and has the same hash")
-
-                    # Returning false indicates that no file was created, so we shouldn't load documents
-                    return False
-            else:
-                if existing_file.file_hash != file_hash:
-                    raise ValueError(
-                        f"File with name: {file_name} already exists in collection: {collection_id}, but has a different hash"
-                    )
-                else:
-                    logging.info(
-                        f"File: {file_name} in collection: {collection_id} already exists, and has the same hash"
-                    )
-                    st.info(f"Skipping load: File: {file_name} in collection: {collection_id} already exists, and has the same hash")                    
-
-                    # Returning false indicates that no file was created, so we shouldn't load documents
-                    return False
-        
-        return True
-    
-    def upload_files(self, uploaded_files, status):
-        root_temp_dir = "temp"
-
-        if not os.path.exists(root_temp_dir):
-            os.makedirs(root_temp_dir)
-
-        # First upload all of the files- this needs to be done before we process them, in case there are inter-dependencies
-        uploaded_file_paths = []
-        for uploaded_file in uploaded_files:
-            with status.empty():
-                with st.container():
-                    st.info(f"Uploading file: {uploaded_file.name}")
-                    # if use_same_upload_dir is set, create a unique path for each file- 
-                    # because files can and will get split into multiples (e.g. Split into CSVs for excel files)
-                    if st.session_state["use_same_upload_dir"]:
-                        file_path = os.path.join(root_temp_dir, uploaded_file.name)
-                    else:
-                        file_path = os.path.join(root_temp_dir, str(uuid.uuid4()), uploaded_file.name)
-
-                    os.makedirs(os.path.dirname(file_path), exist_ok=True)
-                    with open(file_path, "wb") as f:
-                        f.write(uploaded_file.getbuffer())
-                    
-                    uploaded_file_paths.append(file_path)
-
-        return uploaded_file_paths
-
-    def ingest_files(
-        self,
-        uploaded_files,
-        active_collection,
-        collection_id,
-        status,
-        overwrite_existing_files,
-        split_documents,
-        chunk_size,
-        chunk_overlap,
-    ):
-        status.update(
-            label=f"Ingesting files and adding to {active_collection}",
-            state="running",
-        )
-
-        try:
-            # First upload the files to our temp directory
-            uploaded_file_paths = self.upload_files(uploaded_files, status)
-
-            documents_helper = Documents()
-
-            with status.empty():
-                with st.container():
-                    st.info(f"Processing files... please wait.")
-
-                    # Go through each file and see if it already exists, then split and upload the chunks
-                    
-                    for uploaded_file_path in uploaded_file_paths:
-                        file_name = os.path.basename(uploaded_file_path)
-                        # If we're using the same upload dir for everything, we need to make sure the target path is the file, and not the directory
-                        if st.session_state["use_same_upload_dir"]:
-                            file_directory = uploaded_file_path
-                        else:
-                            file_directory = os.path.dirname(uploaded_file_path)
-                        
-                        with open(uploaded_file_path, 'rb') as file:
-                            file_data = file.read()
-
-                        file_hash = calculate_sha256(uploaded_file_path)                        
-                        
-                        # If we're overwriting, delete the file and any chunks associated with it.
-                        # This will return false if we are not OK to create the file and document chunks because the file exists or we can't delete the old one.
-                        clear_to_create = self.clear_to_create_file_and_docs(file_name, file_hash, collection_id, overwrite_existing_files)
-
-                        if clear_to_create:    
-                            documents = load_and_split_documents(
-                                file_directory, split_documents, chunk_size, chunk_overlap
-                            )
-
-                            # Since files can be split into multiples, we need to get a new list of file names
-                            file_names = list(set([d.metadata["filename"] for d in documents]))
-
-                            for file_name in file_names:
-
-                                if len(file_names) > 1:
-                                    # That means we also have to re-run the clear check
-                                    # This whole step should be refactored so it takes place at a higher level or something where we 
-                                    # don't have to split the documents up first
-                                    file_hash = calculate_sha256(uploaded_file_path)                            
-                                    if not self.clear_to_create_file_and_docs(file_name, file_hash, collection_id, overwrite_existing_files):                                        
-                                        continue
-
-                                    st.info(f"The file {file_name} was split into multiple files ({len(file_names)}), so we're creating a new file for each one")
-                                
-                                file = documents_helper.create_file(
-                                    FileModel(
-                                        collection_id,
-                                        user_id=st.session_state["user_id"],
-                                        file_name=file_name,
-                                        file_hash=calculate_sha256(os.path.join(file_directory, file_name)),
-                                        file_data=file_data
-                                    )
-                                )
-
-                                st.info(f"Loading {len(documents)} chunks for {file_name}")
-                            
-                                for document in documents:
-                                    documents_helper.store_document(
-                                        DocumentModel(
-                                            collection_id=collection_id,
-                                            file_id=file.id,
-                                            user_id=st.session_state["user_id"],
-                                            document_text=document.page_content,
-                                            document_name=document.metadata["filename"],
-                                            additional_metadata=document.metadata,
-                                        )
-                                    )
-
-                                self.classify_file(documents_helper, documents, file)                        
-
-                        
-
-                    status.empty()
-                    status.update(
-                        label=f"✅ Document ingestion complete!",
-                        state="complete",
-                        expanded=False,
-                    )
-
-                    uploaded_files.clear()
-
-        except Exception as e:
-            # debugging streamlit
-            # raise e
-            logging.error(e)
-            print(e)
-            status.error(f"Error: {e}", icon="❌")
-            status.update(
-                label=f"Document ingestion failed!",
-                state="error",
-                expanded=False,
-            )
-
-            try:
-                shutil.rmtree("temp")
-                print(f"Deleted temporary directory")
-            except Exception as e:
-                print(f"Error deleting temporary directory: {e}")
-
-    def classify_file(
-        self, documents_helper: Documents, documents, file
-    ):
-        # Use up to the first 10 document chunks to classify this document
-
-        text = "\n".join([d.page_content for d in documents[:10]])
-        file.file_classification = documents[0].metadata.get(
-            "classification", "Document"
-        )
-
-        general_ai_instance: RequestRouter = st.session_state["general_ai"]
-
-        summarize_string = get_prompt(
-            general_ai_instance.assistant_configuration.request_router.model_configuration.llm_type,
-            "CONCISE_SUMMARIZE_TEMPLATE",
-        ).format(text=text)
-
-        file.file_summary = general_ai_instance.llm.predict(summarize_string)        
-
-        documents_helper.update_file_summary_and_class(
-            file.id, file.file_summary, file.file_classification
-        )
-
-        print(
-            f"File summary: {file.file_summary}, classification: {file.file_classification}"
-        )
 
     def refresh_messages_session_state(self, general_ai_instance):
         """Pulls the messages from the token buffer on the AI for the first time, and put them into the session state"""
@@ -568,30 +208,36 @@ class GeneralUI:
 
             self.show_old_messages(general_ai_instance)
 
+            # st.write("")
+            # st.write("")
+            # st.write("")
+            # st.write("")
+            # st.write("")
+            # st.write("")
+
         # Get user input (must be outside of the container)
         prompt = st.chat_input("Enter your message here", key="chat_input")
 
-        if prompt:
+        if prompt:            
             with main_window_container.container():
                 st.chat_message("user", avatar="👤").markdown(prompt)
 
                 with st.chat_message("assistant", avatar="🤖"):
+                    agent_callback_container = st.container().empty()
                     llm_container = st.container().empty()
                     llm_callbacks = []
                     llm_callbacks.append(StreamingOnlyCallbackHandler(llm_container))
 
-                    agent_callbacks = []
-                    if st.session_state["show_llm_thoughts"]:
-                        print("showing agent thoughts")
-                        agent_callback_container = st.container().empty()
-                        agent_callback = StreamlitCallbackHandler(
-                            agent_callback_container,
-                            expand_new_thoughts=True,
-                            collapse_completed_thoughts=True,
-                        )
-                        agent_callbacks.append(agent_callback)
+                    agent_callbacks = []                    
+                    print("showing agent thoughts")                    
+                    agent_callback = StreamlitCallbackHandler(
+                        agent_callback_container,
+                        expand_new_thoughts=True,
+                        collapse_completed_thoughts=True,
+                    )
+                    agent_callbacks.append(agent_callback)
 
-                    collection_id = self.collection_id_from_option(
+                    collection_id = ui_shared.collection_id_from_option(
                         st.session_state["active_collection"],
                         general_ai_instance.interaction_manager.interaction_id,
                     )
@@ -613,64 +259,7 @@ class GeneralUI:
                     print(f"Result: {result}")
 
                     llm_container.markdown(result)
-
-    def ensure_user(self, email):
-        self.user_email = email
-
-        users_helper = Users()
-
-        user = users_helper.get_user_by_email(self.user_email)
-
-        if not user:
-            st.markdown(f"Welcome to Jarvis, {self.user_email}! Let's get you set up.")
-
-            # Create the user by showing them a prompt to enter their name, location, age
-            name = st.text_input("Enter your name")
-            location = st.text_input("Enter your location")
-
-            if (
-                name and location
-            ):  # Check if both name and location inputs are not empty
-                if st.button("Create Your User!"):
-                    user = users_helper.create_user(
-                        email=self.user_email, name=name, location=location, age=999
-                    )
-                else:
-                    return False
-            else:
-                return False
-        else:
-            return True
-
-    def set_user_id_from_email(self):
-        users_helper = Users()
-
-        user = users_helper.get_user_by_email(self.user_email)
-        st.session_state["user_id"] = user.id
-
-        return user.id
-
-    def create_interaction(self, interaction_summary):
-        interactions_helper = Interactions()
-
-        interactions_helper.create_interaction(
-            id=str(uuid.uuid4()),
-            interaction_summary=interaction_summary,
-            user_id=st.session_state["user_id"],
-        )
-
-    def ensure_interaction(self):
-        """Ensures that an interaction exists for the current user"""
-
-        # Only do this if we haven't already done it
-        if (
-            "interaction_ensured" not in st.session_state
-            or not st.session_state["interaction_ensured"]
-        ):
-            if not self.get_interaction_pairs():
-                self.create_interaction("Empty Chat")
-
-            st.session_state["interaction_ensured"] = True
+                    ui_shared.scroll_to_bottom('column')
 
     def load_configuration(self):
         # Load environment variables from the .env file
@@ -692,43 +281,43 @@ class GeneralUI:
 
         st.title("Hey Jarvis 🤖...")
 
+
 if __name__ == "__main__":
     logging.basicConfig(level=logging.INFO)
 
-    streamlit_ui = GeneralUI()
+    try:
+        general_ui = GeneralUI()
 
-    # Always comes first!
-    streamlit_ui.load_configuration()
+        # Always comes first!
+        general_ui.load_configuration()
 
-    streamlit_ui.set_page_config()
+        general_ui.set_page_config()
 
-    # Get the user from the environment variables
-    user_email = os.environ.get("USER_EMAIL", None)
+        # Get the user from the environment variables
+        user_email = os.environ.get("USER_EMAIL", None)
+        general_ui.user_email = user_email
 
-    if not user_email:
-        raise ValueError("USER_EMAIL environment variable not set")
+        if not user_email:
+            raise ValueError("USER_EMAIL environment variable not set")
 
-    if streamlit_ui.ensure_user(user_email):
-        streamlit_ui.set_user_id_from_email()
+        if ui_shared.ensure_user(user_email):
+            ui_shared.set_user_id_from_email(user_email)
+            ui_shared.ensure_interaction()
+            ui_shared.load_interaction_selectbox(general_ui.load_ai)
+            # Set up columns for chat and collections
+            col1, col2 = st.columns([0.65, 0.35])
 
-        streamlit_ui.ensure_interaction()
+            general_ui.load_ai()
+            ui_shared.setup_new_chat_button()
+            general_ui.create_collections_container(col2)
 
-        streamlit_ui.load_interaction_selectbox()
+            ui_shared.select_documents()
 
-        # Set up columns for chat and collections
-        col1, col2 = st.columns([0.65, 0.35])
+            general_ui.handle_chat(col1)
 
-        print("loading ai")
-        streamlit_ui.load_ai()
-
-        print("selecting conversation")
-        streamlit_ui.select_conversation()
-
-        print("creating collections container")
-        streamlit_ui.create_collections_container(col2)
-
-        print("selecting documents")
-        streamlit_ui.select_documents()
-
-        print("handling chat")
-        streamlit_ui.handle_chat(col1)
+            ui_shared.scroll_to_bottom('column')
+    
+    except Exception as e:
+        # This should only be catching a StopException thrown by streamlit, yet I cannot find it for the fucking life of me.
+        # And after wasting 20 minutes of my life on this, I am done.
+        pass
