+++ requirements.txt
@@ -14,6 +14,7 @@ sqlalchemy
 
 # AI
 # Note: You need a C++ compiler to install llama-cpp-python if you are building this outside of the docker container
+# https://cmake.org/download/#latest
 llama-cpp-python
 openai
 #openai-functions
@@ -38,8 +39,9 @@ langchain
 # Maffs
 # torch
 # To install torch for GPU (CUDA 12.1), use the following command:
-# pip3 install --force-reinstall --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cu121
-# ^^ The force reinstall is needed because the version of torch that gets installed as a dependency in other places is not correct
+# pip3 install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cu121
+# Force reinstall might be needed because the version of torch that gets installed as a dependency in other places is not correct
+
 # NOTE: The pip install of ffmpeg-python is broken, so you need to install ffmpeg manually
 # After you install ffmpeg (download from the website), you need to add the bin folder to your path
 # ffmpeg
