+++ src/ai/rag_ai.py
@@ -17,7 +17,8 @@ from src.configuration.assistant_configuration import (
 )
 
 from src.ai.destinations.output_parser import (
-    CustomStructuredChatOutputParserWithRetries,
+    CustomStructuredChatOutputParserWithRetries,   
+    CustomOutputFixingParser 
 )
 
 from src.ai.interactions.interaction_manager import InteractionManager
@@ -102,7 +103,7 @@ class RetrievalAugmentedGenerationAI:
                     self.configuration.model_configuration.llm_type,
                     "TOOLS_FORMAT_INSTRUCTIONS",
                 ),
-                "output_parser": CustomStructuredChatOutputParserWithRetries(),
+                "output_parser": CustomStructuredChatOutputParserWithRetries.from_llm(llm=self.llm), #(output_fixing_parser=CustomOutputFixingParser()),
                 "input_variables": [
                     "input",
                     "loaded_documents",
